Rounding up of scores to display only upto 2 decimal points

+++ backend/src/main/java/trustworthy/analyser/dataObjects/AnalyserResponseObject.java
@@ -11,6 +11,6 @@ public class AnalyserResponseObject {
     private double availabilityScore;
     private double resilienceScore;
     private double reliabilityScore;
-    private double trustworthyScore;
+    private int trustworthyScore;
     private String verdict;
 }
